rename app to app.jsx
